fix: на ревью Сергею

+++ src/test/java/ru/yandex/practicum/filmorate/utilities/FilmRecommendationTest.java
@@ -10,8 +10,8 @@ class FilmRecommendationTest {
 
     @Test
     void getRecommendationTest_standardDeviation() {
-        Map<Long, Double> targetUserRates = initializeTargetUserData();
-        Map<Long, Map<Long, Double>> similarUsersFilmsRates = initializeSimilarUsersData();
+        Map<Long, Integer> targetUserRates = initializeTargetUserData();
+        Map<Long, Map<Long, Integer>> similarUsersFilmsRates = initializeSimilarUsersData();
 
         List<Long> expected = List.of(4L, 5L, 6L);
         List<Long> calculated = FilmRecommendation.getRecommendation(targetUserRates, similarUsersFilmsRates);
@@ -20,8 +20,8 @@ class FilmRecommendationTest {
 
     @Test
     void getRecommendationTest_targetUserDontLikeAnyFilm() {
-        Map<Long, Double> targetUserRates = new HashMap<>();
-        Map<Long, Map<Long, Double>> similarUsersFilmsRates = initializeSimilarUsersData();
+        Map<Long, Integer> targetUserRates = new HashMap<>();
+        Map<Long, Map<Long, Integer>> similarUsersFilmsRates = initializeSimilarUsersData();
 
         List<Long> expected = List.of(1L, 2L, 3L, 4L, 5L, 6L);
         List<Long> calculated = FilmRecommendation.getRecommendation(targetUserRates, similarUsersFilmsRates);
@@ -30,8 +30,8 @@ class FilmRecommendationTest {
 
     @Test
     void getRecommendationTest_targetUserHasNoIntersections() {
-        Map<Long, Double> targetUserRates = initializeTargetUserData();
-        Map<Long, Map<Long, Double>> similarUsersFilmsRates = new HashMap<>();
+        Map<Long, Integer> targetUserRates = initializeTargetUserData();
+        Map<Long, Map<Long, Integer>> similarUsersFilmsRates = new HashMap<>();
         similarUsersFilmsRates.put(1L, new HashMap<>());
         similarUsersFilmsRates.put(2L, new HashMap<>());
         similarUsersFilmsRates.put(3L, new HashMap<>());
@@ -41,33 +41,33 @@ class FilmRecommendationTest {
         Assertions.assertEquals(expected, calculated);
     }
 
-    private Map<Long, Double> initializeTargetUserData() {
-        Map<Long, Double> targetUserRates = new HashMap<>();
-        targetUserRates.put(1L, 1.0);
-        targetUserRates.put(2L, 1.0);
-        targetUserRates.put(3L, 1.0);
-        targetUserRates.put(7L, 1.0);
+    private Map<Long, Integer> initializeTargetUserData() {
+        Map<Long, Integer> targetUserRates = new HashMap<>();
+        targetUserRates.put(1L, 1);
+        targetUserRates.put(2L, 1);
+        targetUserRates.put(3L, 1);
+        targetUserRates.put(7L, 1);
         return targetUserRates;
     }
 
-    private Map<Long, Map<Long, Double>> initializeSimilarUsersData() {
-        Map<Long, Map<Long, Double>> similarUsersFilmsRates = new HashMap<>();
+    private Map<Long, Map<Long, Integer>> initializeSimilarUsersData() {
+        Map<Long, Map<Long, Integer>> similarUsersFilmsRates = new HashMap<>();
 
-        Map<Long, Double> firstSimilarUserRates = new HashMap<>();
-        firstSimilarUserRates.put(1L, 1.0);
-        firstSimilarUserRates.put(2L, 1.0);
-        firstSimilarUserRates.put(4L, 8.0);
+        Map<Long, Integer> firstSimilarUserRates = new HashMap<>();
+        firstSimilarUserRates.put(1L, 1);
+        firstSimilarUserRates.put(2L, 1);
+        firstSimilarUserRates.put(4L, 8);
 
-        Map<Long, Double> secondSimilarUserRates = new HashMap<>();
-        secondSimilarUserRates.put(2L, 1.0);
-        secondSimilarUserRates.put(3L, 1.0);
-        secondSimilarUserRates.put(5L, 1.0);
+        Map<Long, Integer> secondSimilarUserRates = new HashMap<>();
+        secondSimilarUserRates.put(2L, 1);
+        secondSimilarUserRates.put(3L, 1);
+        secondSimilarUserRates.put(5L, 1);
 
-        Map<Long, Double> thirdSimilarUserRates = new HashMap<>();
-        thirdSimilarUserRates.put(1L, 1.0);
-        thirdSimilarUserRates.put(2L, 1.0);
-        thirdSimilarUserRates.put(3L, 1.0);
-        thirdSimilarUserRates.put(6L, 1.0);
+        Map<Long, Integer> thirdSimilarUserRates = new HashMap<>();
+        thirdSimilarUserRates.put(1L, 1);
+        thirdSimilarUserRates.put(2L, 1);
+        thirdSimilarUserRates.put(3L, 1);
+        thirdSimilarUserRates.put(6L, 1);
 
         similarUsersFilmsRates.put(1L, firstSimilarUserRates);
         similarUsersFilmsRates.put(2L, secondSimilarUserRates);
